+++ packages/cli/bin/cli.js
@@ -1,28 +1,35 @@
 #!/usr/bin/env -S node --no-warnings
 
 import cli from '#src/main.js'
+import config from '#src/lib/conf/config.js'
 import packageConfig from '#root/package.json' assert { type: 'json' }
 import updateNotifier from 'update-notifier'
 
 process.removeAllListeners('warning')
 
+/**
+ * Interval constants in milliseconds
+ */
 const INTERVAL = Object.freeze({
-  MINUTES: 1000 * 60 * 1,
+  MINUTE: 1000 * 60 * 1,
   HOURLY: 1000 * 60 * 60,
   DAILY: 1000 * 60 * 60 * 24,
   WEEKLY: 1000 * 60 * 60 * 24 * 7
 })
 
+const updateChannel = config.get('updateChannel')
+const updateIterval = config.get('updateIterval')
+
 /**
- * Check for quire-cli updates
+ * Create a notifier to Check for quire-cli updates
  * @see https://github.com/yeoman/update-notifier#usage
  *
- * @todo user configuration of choosen update interval
+ * @todo refactor to check multiple channels
  */
 const notifier = updateNotifier({
-  distTag: 'latest',
+  distTag: updateChannel,
   pkg: packageConfig,
-  updateCheckInterval: INTERVAL.DAILY,
+  updateCheckInterval: INTERVAL[updateIterval],
 })
 
 notifier.notify({ defer: false })