+++ packages/cli/CHANGELOG.md
@@ -12,6 +12,13 @@ Changelog entries are classified using the following labels:
 - `Fixed`: for any bug fixes
 - `Removed`: for deprecated features removed in this release
 
+## [1.0.0-rc.18]
+
+### Added
+
+- Quire CLI configuration management using the `conf` package
+- `quire conf` command to manage cli configuration properties
+
 ## [1.0.0-rc.17]
 
 ### Bumped